Made Generic Wizard for reuse by all
