+++ inventory/views/__init__.py
@@ -3,4 +3,6 @@ from .connection_autocomplete import ConnectionAutocomplete
 from .disposition_autocomplete import DispositionAutocomplete
 from .tag_autocomplete import TagAutocomplete
 from .items_list_view import ItemsListView
+from .generic_wizard import GenericWizard
 from .make_item_wizard import MakeItemWizard
+from .bulk_image_upload import BulkImageUpload
