+++ inventory/views/generic_wizard.py
@@ -1,0 +1,98 @@
+from django.views.generic import View
+from django.views.decorators.cache import never_cache
+from django.utils.decorators import method_decorator
+from django.contrib.auth.decorators import login_required
+from django.http import Http404
+from django.http import HttpResponseRedirect
+from django.urls import reverse
+from django.shortcuts import render
+from django.contrib import messages
+
+
+class GenericWizard(View):
+    ##############
+    #  This is an abstract class, it gives the logic for rolling through
+    #  a set of forms as a wizard, to use it:
+    #     - instantiate form_sets = a dict of integers (-1 to however many)
+    #          - with a sub-dict with a "the_form", "next_form", "next_title"
+    #          - there must be a -1 with the_form = None
+    #          - there must be a last item with next_form and next_title 
+    #            as None
+    #     - create setup_form - which can make any form in the set, the first
+    #          form can be  made via either get or post, all forms after that
+    #          are submitted as posts.
+    #     - finish_valid_form - what to do when a form is deemed valid
+    #     - finish - place to put any messaging and return a URL for how to 
+    #          return to a main spot
+    ##############
+    step = -1
+    max = 1
+
+    def groundwork(self, request, args, kwargs):
+        self.step = int(request.POST.get("step", -1))
+        self.return_url = reverse('items_list', urlconf='inventory.urls')
+
+    def make_context(self, request):
+        context = {
+            'page_title': self.page_title,
+            'title': self.page_title,
+            'subtitle': self.current_form_set['next_title'],
+            'forms': [self.form],
+            'first': self.current_form_set['the_form'] is None,
+            'last': self.form_sets[self.step+1]['next_form'] is None,
+        }
+        return context
+
+    def make_back_forms(self, request):
+        self.step = self.step - 2
+        self.current_form_set = self.form_sets[self.step]
+
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(GenericWizard, self).dispatch(*args, **kwargs)
+
+    @never_cache
+    def get(self, request, *args, **kwargs):
+        redirect = self.groundwork(request, args, kwargs)
+        self.current_form_set = self.form_sets[-1]
+        self.form = self.setup_form(self.current_form_set['next_form'])
+        return render(request, self.template, self.make_context(request))
+
+    @never_cache
+    @method_decorator(login_required)
+    def post(self, request, *args, **kwargs):
+        self.groundwork(request, args, kwargs)
+        if 'cancel' in list(request.POST.keys()):
+            messages.success(request, "The last update was canceled.")
+            return HttpResponseRedirect(self.return_url)
+
+        if 'next' in list(request.POST.keys()) or (
+                'finish' in list(request.POST.keys())):
+            self.current_form_set = self.form_sets[self.step]
+            self.form = self.setup_form(
+                self.current_form_set['the_form'],
+                request.POST)
+            if not self.form.is_valid():
+                self.step = self.step - 1
+                self.current_form_set = self.form_sets[self.step]
+                return render(request, self.template, self.make_context(
+                    request))
+            self.finish_valid_form()
+            if 'finish' in list(request.POST.keys()):
+                return HttpResponseRedirect(self.finish(request))
+        elif 'back' in list(request.POST.keys()):
+            self.make_back_forms(request)
+        else:
+            messages.error(
+                request,
+                "Button Click Unclear.  If you did not tamper with the form," +
+                " contact us.")
+            self.current_form_set = {'next_form': None}
+
+        if self.current_form_set['next_form'] is not None:
+            self.form = self.setup_form(self.current_form_set['next_form'])
+            return render(request, self.template, self.make_context(request))
+
+        messages.error(request, "Unexpected logic flow.  Contact support.")
+
+        return HttpResponseRedirect(self.return_url)
